Use file monitor in ServerSpec
This test periodically fails on travis. To fix it, I bumped the timeout
to 3 minutes. I also replaced the file system poll with a
FileEventMonitor so that increasing the timeout should have no effect
when run locally and the test will proceed almost instantaneously when
the file is created. I also added a timeout after the file is detected
to ensure that the server is really ready to receive requests.