Bump io
This new version of io breaks source and binary compatibility for any
class that uses TypedPath. We haven't released a version of sbt with the
old TypedPath apis, so that is fine.

I also noticed a number of places where we were calling .asFile on a
file. This is redundant because asFile is an extension method on File
that just returns the underlying file.